:sparkles: add CoinGecko API integration for fetching coin mappings
