+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod coingecko;
 pub(crate) mod config;
 pub mod constants;
 pub mod models;
